can read and write strategy flow

+++ database/src/strategy.ts
@@ -5,46 +5,50 @@ import {
   CreateStrategy,
   DeleteStrategy,
   ReadStrategy,
+  ReadStrategyAccountId,
   RenameStrategy,
   Strategy,
   StrategyKey,
   createdNow,
   deletedNow,
-  updatedNow,
-  normalizeStrategyName,
+  isAccountKey,
   isStrategyName,
+  normalizeStrategyName,
+  updatedNow,
 } from "@ggbot2/models";
 import { v4 as uuidv4 } from "uuid";
 import {
   readAccountStrategyList,
   writeAccountStrategyList,
 } from "./accountStrategyList.js";
+import {
+  ErrorInvalidStrategyName,
+  ErrorMissingAccountId,
+  ErrorPermissionDeniedCannotDeleteStrategy,
+  ErrorStrategyNotFound,
+} from "./errors.js";
+import { strategyKeyToDirname } from "./strategyKey.js";
 
 export const strategyDirnamePrefix = () => "strategy";
 
 export const strategyDirname = (strategyKey: StrategyKey) =>
   `${strategyDirnamePrefix()}/${strategyKeyToDirname(strategyKey)}`;
 
-export const strategyKeyToDirname = ({
-  strategyId,
-  strategyKind,
-}: StrategyKey) => `strategyKind=${strategyKind}/strategyId=${strategyId}`;
-
 export const strategyPathname = (strategyKey: StrategyKey) =>
   `${strategyDirname(strategyKey)}/strategy.json`;
 
 export const copyStrategy: CopyStrategy["func"] = async ({
   accountId,
-  strategyKind,
-  strategyId,
   name,
+  ...strategyKey
 }) => {
-  const strategy = await readStrategy({ strategyKind, strategyId });
-  if (!strategy)
-    throw new Error(
-      `Strategy not found, kind=${strategyKind} id=${strategyId}`
-    );
-  return await createStrategy({ accountId, kind: strategyKind, name });
+  const strategy = await readStrategy(strategyKey);
+  if (!strategy) throw new ErrorStrategyNotFound(strategyKey);
+  return await createStrategy({
+    accountId,
+    kind: strategyKey.strategyKind,
+    name,
+  });
 };
 
 export const createStrategy: CreateStrategy["func"] = async ({
@@ -54,7 +58,7 @@ export const createStrategy: CreateStrategy["func"] = async ({
 }) => {
   const strategyId = uuidv4();
   const strategyKind = kind;
-  const strategyKey: StrategyKey = { strategyId, strategyKind };
+  const strategyKey = { strategyId, strategyKind };
   const data: Strategy = {
     id: strategyId,
     kind,
@@ -84,37 +88,37 @@ export const readStrategy: ReadStrategy["func"] = async (strategyKey) => {
   return data as Strategy;
 };
 
+export const readStrategyAccountId: ReadStrategyAccountId["func"] = async (
+  strategyKey
+) => {
+  const Key = strategyPathname(strategyKey);
+  const data = await getObject({ Key });
+  if (!data) throw new ErrorStrategyNotFound(strategyKey);
+  if (!isAccountKey(data)) throw new ErrorMissingAccountId();
+  return data.accountId;
+};
+
 export const renameStrategy: RenameStrategy["func"] = async ({
   accountId,
-  strategyId,
-  strategyKind,
   name,
+  ...strategyKey
 }) => {
-  if (!isStrategyName(name))
-    throw new TypeError(`Invalid strategy name ${name}`);
-  const strategyKey: StrategyKey = { strategyId, strategyKind };
+  if (!isStrategyName(name)) throw new ErrorInvalidStrategyName(name);
   const strategy = await readStrategy(strategyKey);
-  if (!strategy)
-    throw new Error(
-      `Strategy not found, kind=${strategyKind} id=${strategyId}`
-    );
-  if (strategy.accountId !== accountId)
-    throw new Error(
-      `Permission denied, accountId=${accountId} cannot rename strategy kind=${strategyKind} id=${strategyId}`
-    );
-  const renamedStrategy = { ...strategy, name: normalizeStrategyName(name) };
-  const Key = strategyPathname(strategyKey);
-  await putObject({ Key, data: renamedStrategy });
+  if (!strategy) throw new ErrorStrategyNotFound(strategyKey);
+  if (strategy.accountId === accountId) {
+    const renamedStrategy = { ...strategy, name: normalizeStrategyName(name) };
+    const Key = strategyPathname(strategyKey);
+    await putObject({ Key, data: renamedStrategy });
+  }
   const strategies = (await readAccountStrategyList({ accountId })) ?? [];
   await writeAccountStrategyList({
     accountId,
-    strategies: strategies.map(
-      ({ strategyId: itemStrategyId, name: itemName, ...item }) => ({
-        strategyId: itemStrategyId,
-        name: strategyId === itemStrategyId ? name : itemName,
-        ...item,
-      })
-    ),
+    strategies: strategies.map(({ strategyId, name: itemName, ...item }) => ({
+      strategyId: strategyKey.strategyId,
+      name: strategyId === strategyKey.strategyId ? name : itemName,
+      ...item,
+    })),
   });
   return updatedNow();
 };
@@ -123,6 +127,12 @@ export const deleteStrategy: DeleteStrategy["func"] = async ({
   accountId,
   ...strategyKey
 }) => {
+  const ownerId = await readStrategyAccountId(strategyKey);
+  if (accountId !== ownerId)
+    throw new ErrorPermissionDeniedCannotDeleteStrategy({
+      accountId,
+      strategyKey,
+    });
   const Key = strategyPathname(strategyKey);
   await deleteObject({ Key });
   const strategies = (await readAccountStrategyList({ accountId })) ?? [];
